maybe_async_cfg

Attribute Macro maybe

#[maybe]

Marks the code that can be presented in several versions.

The maybe macro has the following parameters:

disable

The macro with disable parameter will do nothing, like noop. Use it to write and debug initial async code.

prefix

The name of maybe-async-cfg crate. If not set, "maybe_async_cfg" will be used.

sync, async

Defines versions of the code: the item to which the attribute maybe refers will be replaced with multiple copies (one for each version), which will be modified according to the version kind and its parameters.

For the sync version, the item will be converted from async to sync code by deleting the async and await keywords. The types Future<Output=XXX> will also be replaced with just XXX. For the async version, the item will be left async.

In any case, the item will be converted according to all the parameters described below. For functions, structs/enums and traits, the name will be changed as if it is mentioned in the idents list (if it is not explicitly specified there and if keep_self is not present).

All other parameters will be passed to all versions (with merging).

Therefore, those parts of the version parameters that match in all versions can be specified here. For example, this is the expected behavior for the idents list.

Every version has the following parameters:

disable

Ignore this version entirely.

key

Defines unique name of the version to use it in only_if/remove_if conditions. If omitted, sync/async will be used.

#[maybe_async_cfg::maybe(
    sync(key="foo", feature="use_sync"),
    async(key="bar", feature="use_async"),
)]
struct Struct {
    f: usize,
     
    // This field will only be present in sync version
    #[maybe_async_cfg::only_if(key="foo")]
    sync_only_field: bool,
}

After conversion:

#[cfg(feature="use_sync")]
struct StructSync {
    f: usize,
    sync_only_field: bool,
}
#[cfg(feature="use_async")]
struct StructAsync {
    f: usize,
}

cfg

Defines the condition (based on features), under which the current version should appear.

Note: conditions like feature = "...", not(...), all(...), any(...) will be processed correctly, even if the cfg(...) was omitted.

#[maybe_async_cfg::maybe(
    sync(cfg(feature="use_sync")),
    async(feature="use_async")
)]
struct Struct {
    f: Foo,
}

After conversion:

#[cfg(feature="use_sync")]
struct StructSync {
    f: Foo,
}
#[cfg(feature="use_async")]
struct StructAsync {
    f: Foo,
}

idents

Defines a list of identifiers that should be renamed depending on the version of the code.

Each identifier can have the following clarifying parameters:

snake, fn, mod

means that this is the snake-case name of the function or module and it should be converted by adding the suffixes "_sync"/"_async" (otherwise, the suffixes "Sync"/"Async" will be used).

use

in use lists, using this identifier will result in renaming via the as expression, rather than a simple replacement as is. In other cases, a simple replacement will be used.

keep

this identifier will not be converted anywhere

sync, async

specifies the name that will be used in the corresponding version of the code. Overrides the standard scheme of suffixes used by default. If the parameter value is omitted, the identifier will not be renamed in this case.

#[maybe_async_cfg::maybe(
    idents(
        Foo,
        Bar,
        baz(fn),
        Qux(use),
        waldo(sync, async="async_waldo"),
        xyzzy(fn, use, sync="xizzy_the_sync_func"),
    ),
    sync(feature="use_sync"),
    async(feature="use_async"),
)]
async fn func() {
    struct Foo {}
    use extcrate::{
        Bar,
        baz,
        Qux,
        waldo::{
            plugh,
            xyzzy
        }
    };
    let _ = baz( Foo {}, Bar::new() ).await;
    let _ = xizzy( Qux::flob(b).await );
}

After conversion:

#[cfg(feature="use_sync")]
fn func_sync() {
    struct FooSync {}
    use extcrate::{
        BarSync,
        baz_sync,
        Qux as QuxSync,
        waldo::{
            plugh,
            xyzzy as xizzy_the_sync_func
        }
    };         
    let _ = baz_sync( FooSync {}, BarSync::new() );
    let _ = xizzy_the_sync_func( QuxSync::flob() );
}
#[cfg(feature="use_async")]
async fn func_async() {
    struct FooAsync {}
    use extcrate::{
        BarAsync,
        baz_async,
        Qux as QuxAsync,
        async_waldo::{
            plugh,
            xyzzy as xyzzy_async
        }
    };
    let _ = baz_async( FooAsync {}, BarAsync::new() ).await;
    let _ = xyzzy_async( QuxAsync::flob().await );     
}

keep_self

Do not change name of item to which attribute maybe refers.

self

Defines the name that will be assigned to the item in this variant.

send

If send = "Send" or send = "true" is present, the attribute #[async_trait::async_trait] will be added before the async code. If send = "?Send" or send = "false" then #[async_trait::async_trait(?Send)] will be added.

drop_attrs

Remove any attributes with specified names.

#[maybe_async_cfg::maybe(
    sync(feature="use_sync", drop_attrs(attr)),
    async(feature="use_async"),
)]
struct Struct {
    f: usize,

    // This attribute will be removed in sync version
    #[attr(param)]
    field1: bool,
}

After conversion:

#[cfg(feature="use_sync")]
struct StructSync {
    f: usize,
    field1: bool,
}
#[cfg(feature="use_async")]
struct StructAsync {
    f: usize,
    #[attr(param)]
    field1: bool,
}

replace_features

Replace one feature name with another.

#[maybe_async_cfg::maybe(
    sync(feature="use_sync", replace_feature("secure", "secure_sync")),
    async(feature="use_async"),
)]
struct Struct {
    f: usize,
    // In sync version "secure" feature will be replaced with "secure_sync" feature
    #[cfg(feature="secure")]
    field: bool,
}

After conversion:

#[cfg(feature="use_sync")]
struct StructSync {
    f: usize,
    #[cfg(feature="secure_sync")]
    field: bool,
}
#[cfg(feature="use_async")]
struct StructAsync {
    f: usize,
    #[cfg(feature="secure")]
    field: bool,
}

inner, outer

Adds some attributes to the generated code. Inner attributes will appear below attribute #[cfg(...)], outer attributes will appear above it.

Note: if the version parameter is not parsed as a parameter of some other type, it will be interpreted as an inner attribute.

Useful for testing: just write test in version parameters.

#[maybe_async_cfg::maybe(
    sync(feature="secure_sync", test, "resource(path = \"/foo/bar\")", outer(xizzy)),
    async(feature="secure_sync", inner(baz(qux), async_attributes::test)),
)]
async fn test_func() {
    todo!()
}

After conversion:

#[xizzy]
#[cfg(feature="use_sync")]
#[test]
#[resource(path = "/foo/bar")]
fn test_func_sync() {
    todo!()
}
#[cfg(feature="use_async")]
#[baz(qux)]
#[async_attributes::test]
async fn test_func_async() {
    todo!()
}

In other cases, the following rules apply:

name-value pairs (xxx = "yyy") with a name other than key, prefix, send and feature will produce an error.

feature = "...", not(...), all(...), any(...) will be interpreted as condition for current version (as wrapped in cfg(...)).

all another parameters will be interpreted as inner attribute for current version (as wrapped in inner(...)).

Formal syntax

ParametersList :
Parameter (, Parameter)*

Parameter :
disable
| keep_self
| prefix = STRING_LITERAL
| (sync | async) ( VersionParametersList )
| idents ( IdentsList )\

VersionParametersList :
VersionParameter (, VersionParameter)*

VersionParameter :
disable
| keep_self
| key = STRING_LITERAL
| feature = STRING_LITERAL
| self = STRING_LITERAL
| send = ("" | "Send" | "true" | "?Send" | "false")
| (cfg | any | all | not) ( ANY_CFG_CONDITION )
| idents ( IdentsList )
| (outer | inner) ( AttributesList )
| replace_feature ( STRING_LITERAL , STRING_LITERAL )
| drop_attrs ( IdentifiersList )
| Attribute

Path :
IDENTIFIER (:: IDENTIFIER)+

IdentifiersList :
IDENTIFIER (, IDENTIFIER)*

IdentsList :
Ident (, Ident)*

Ident :
IDENTIFIER (( IdentParametersList ))?

IdentParametersList :
IdentParameter (, IdentParameter)*

IdentParameter :
keep
| use
| (snake | fn | mod )
| use
| (sync | async | IDENTIFIER) (= STRING_LITERAL)?

AttributesList :
Attribute (, Attribute)*

Attribute :
(IDENTIFIER | Path) (( ANY_VALID_ARGS ))?
| STRING_LITERAL